:recycle: refactor: name of core tool generics
