+++ packages/core/deep.merge.ts
@@ -1,10 +1,12 @@
-import type { NestedObject } from "./nested.object.type"
+import type { NestedObject } from "../utils"
 
 /**
  * @param nestedObject string key object
  * @returns Merge nested style object deeply
  */
-const deepMerge = <T>(...nestedObject: T[]): T =>
+const deepMerge = <MergeTargetObject>(
+    ...nestedObject: MergeTargetObject[]
+): MergeTargetObject =>
     nestedObject.reduce<NestedObject>((mergedObject, currentObject) => {
         for (const [currKey, currValue] of Object.entries(
             currentObject as NestedObject
@@ -16,19 +18,15 @@ const deepMerge = <T>(...nestedObject: T[]): T =>
                     mergedObject[currKey] = currValue
                 } else {
                     if (currValue) {
-                        if (typeof mergedObject[currKey] === "string") {
-                            deepMerge(mergedObject[currKey], currValue)
-                        } else {
-                            mergedObject[currKey] = deepMerge(
-                                mergedObject[currKey],
-                                currValue
-                            )
-                        }
+                        mergedObject[currKey] = deepMerge(
+                            mergedObject[currKey],
+                            currValue
+                        )
                     }
                 }
             }
         }
         return mergedObject
-    }, {}) as T
+    }, {}) as MergeTargetObject
 
 export { deepMerge }
