+++ packages/core/flatten.object.ts
@@ -1,10 +1,12 @@
-import type { NestedObject } from "./nested.object.type"
+import type { NestedObject } from "../utils"
 
 /**
  * @param object string key object
  * @returns Flatten nested object into `string[]`
  */
-const flattenObject = <T extends NestedObject>(object: T): string[] =>
+const flattenObject = <FlattenTargetObject extends NestedObject>(
+    object: FlattenTargetObject
+): string[] =>
     Object.values(object ?? {})
         .map((value) =>
             typeof value !== "string"
