+++ packages/core/get.tailwind.class.ts
@@ -1,12 +1,13 @@
+import type { NestedObject } from "../utils"
 import { flattenObject } from "./flatten.object"
 import { getStyleClass } from "./get.styleclass"
-import type { NestedObject } from "./nested.object.type"
 
 /**
  * @param styleObject
  * @returns tailwind class
  */
-const getTailwindClass = <T>(styleObject: T): string =>
-    getStyleClass(flattenObject(styleObject as NestedObject))
+const getTailwindClass = <TailwindStyleObject>(
+    styleObject: TailwindStyleObject
+): string => getStyleClass(flattenObject(styleObject as NestedObject))
 
 export { getTailwindClass }
